A bit more GLText refactoring
Moved code from GlWindow.cpp, renamed local variables etc.
